requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary
python-dotenv==1.0.1
docker
jinja2
GitPython==3.1.43
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]
httpx